+++ lib/usdKatana/utils.cpp
@@ -58,6 +58,7 @@
 #include <pxr/usd/usd/collectionAPI.h>
 #include <pxr/usd/usd/modelAPI.h>
 #include <pxr/usd/usd/prim.h>
+#include <pxr/usd/usd/primRange.h>
 #include <pxr/usd/usd/relationship.h>
 #include <pxr/usd/usdGeom/boundable.h>
 #include <pxr/usd/usdGeom/camera.h>
@@ -1015,8 +1016,9 @@ std::string UsdKatanaUtils::GenerateShadingNodeHandle(const UsdPrim& shadingNode
     return name;
 }
 
-void
-_FindCameraPaths_Traversal( const UsdPrim &prim, SdfPathVector *result )
+void _FindCameraPaths_Traversal(const UsdPrim& prim,
+                                const bool traverseModelHierarchyOnly,
+                                SdfPathVector* const result)
 {
     // Recursively traverse model hierarchy for camera prims.
     // Note 1: this requires that either prim types be lofted above
@@ -1025,11 +1027,6 @@ _FindCameraPaths_Traversal( const UsdPrim &prim, SdfPathVector *result )
     //         We have made this restriction consciously to reduce the
     //         latency of camera-enumeration
 
-    // If set, this allows for better traversal for global attributes (camera list and light lists)
-    // by utilizing USD Prim children filters to check for prims in the model hierarchy only,
-    // rather than the default Prim child traversal.
-    static const bool traverseModelHierarchyOnly =
-        TfGetenvBool("KATANA_USD_GLOBALS_TRAVERSE_MODEL_HIERARCHY", true);
     auto flags = UsdPrimDefaultPredicate;
     if (traverseModelHierarchyOnly)
     {
@@ -1039,24 +1036,25 @@ _FindCameraPaths_Traversal( const UsdPrim &prim, SdfPathVector *result )
         if (child->IsA<UsdGeomCamera>()) {
             result->push_back(child->GetPath());
         }
-        _FindCameraPaths_Traversal(*child, result);
+        _FindCameraPaths_Traversal(*child, traverseModelHierarchyOnly, result);
     }
 }
 
-SdfPathVector UsdKatanaUtils::FindCameraPaths(const UsdStageRefPtr& stage)
+SdfPathVector UsdKatanaUtils::FindCameraPaths(const UsdStageRefPtr& stage,
+                                              const bool traverseModelHierarchyOnly)
 {
     SdfPathVector result;
-    _FindCameraPaths_Traversal( stage->GetPseudoRoot(), &result );
+    _FindCameraPaths_Traversal(stage->GetPseudoRoot(), traverseModelHierarchyOnly, &result);
     return result;
 }
 
 // This works like UsdLuxListAPI::ComputeLightList() except it tries to
 // maintain the order discovered during traversal.
-static void
-_Traverse(const UsdPrim &prim,
-          UsdLuxListAPI::ComputeMode mode,
-          std::set<SdfPath, SdfPath::FastLessThan> &seen,
-          SdfPathVector *lights)
+static void _Traverse(const UsdPrim& prim,
+                      const UsdLuxListAPI::ComputeMode mode,
+                      const bool traverseModelHierarchyOnly,
+                      std::set<SdfPath, SdfPath::FastLessThan>& seen,
+                      SdfPathVector* const lights)
 {
     if (!prim)
         return;
@@ -1092,12 +1090,6 @@ _Traverse(const UsdPrim &prim,
     }
     // Traverse descendants.
     auto flags = UsdPrimIsActive && !UsdPrimIsAbstract && UsdPrimIsDefined;
-
-    // If set, this allows for better traversal for global attributes (camera list and light lists)
-    // by utilizing USD Prim children filters to check for prims in the model hierarchy only,
-    // rather than the default Prim child traversal.
-    static const bool traverseModelHierarchyOnly =
-        TfGetenvBool("KATANA_USD_GLOBALS_TRAVERSE_MODEL_HIERARCHY", true);
     if (traverseModelHierarchyOnly && mode == UsdLuxListAPI::ComputeModeConsultModelHierarchyCache)
     {
         // When consulting the cache we only traverse model hierarchy.
@@ -1109,11 +1101,12 @@ _Traverse(const UsdPrim &prim,
     }
     for (const UsdPrim &child:
              prim.GetFilteredChildren(UsdTraverseInstanceProxies(flags))) {
-        _Traverse(child, mode, seen, lights);
+        _Traverse(child, mode, traverseModelHierarchyOnly, seen, lights);
     }
 }
 
-SdfPathVector UsdKatanaUtils::FindLightPaths(const UsdStageRefPtr& stage)
+SdfPathVector UsdKatanaUtils::FindLightPaths(const UsdStageRefPtr& stage,
+                                             const bool traverseModelHierarchyOnly)
 {
 /* XXX -- ComputeLightList() doesn't try to maintain an order.  That
           should be okay for lights but it does cause differences in
@@ -1134,8 +1127,11 @@ SdfPathVector UsdKatanaUtils::FindLightPaths(const UsdStageRefPtr& stage)
     SdfPathVector result;
     std::set<SdfPath, SdfPath::FastLessThan> seen;
     for (const auto &child: stage->GetPseudoRoot().GetChildren()) {
-        _Traverse(child, UsdLuxListAPI::ComputeModeConsultModelHierarchyCache,
-                  seen, &result);
+        _Traverse(child,
+                  UsdLuxListAPI::ComputeModeConsultModelHierarchyCache,
+                  traverseModelHierarchyOnly,
+                  seen,
+                  &result);
     }
     return result;
 }
@@ -1665,7 +1661,10 @@ void UsdKatanaUtils::ShaderToAttrsBySdr(const UsdPrim& prim,
     auto rendererItr = s_contextNameToRenderer.find(shaderPrefix);
     shaderPrefix =
         rendererItr != s_contextNameToRenderer.end() ? rendererItr->second : shaderPrefix;
-    const std::string shaderContextCased = TfStringCapitalize(shaderContext);
+    const std::string shaderContextCased =
+        (shaderPrefix == "prman" && shaderContext == "lightFilter")
+            ? "Lightfilter"
+            : TfStringCapitalize(shaderContext);
     attrs.set(shaderPrefix + shaderContextCased + "Shader", FnKat::StringAttribute(shaderId));
     attrs.set(shaderPrefix + shaderContextCased + "Params", shaderBuilder.build());
 }
@@ -2291,21 +2290,36 @@ bool UsdKatanaUtilsLightListAccess::SetLinks(const UsdCollectionAPI& collectionA
         // so assume that we do.
         isLinked = true;
     }
-    else {
-        UsdCollectionAPI::MembershipQuery query =
-            collectionAPI.ComputeMembershipQuery();
-        UsdCollectionAPI::MembershipQuery::PathExpansionRuleMap linkMap =
-            query.GetAsPathExpansionRuleMap();
-        for (const auto &entry: linkMap) {
-            if (entry.first == SdfPath::AbsoluteRootPath()) {
+    if (!isLinked)
+    {
+        collectionAPI.GetIncludeRootAttr().Get(&isLinked);
+    }
+    const UsdCollectionAPI::MembershipQuery query = collectionAPI.ComputeMembershipQuery();
+    if (!isLinked && query.HasExpression())
+    {
+        const SdfPathSet includedPaths{
+            UsdCollectionAPI::ComputeIncludedPaths(query, _usdInArgs->GetStage())};
+        std::transform(includedPaths.begin(),
+                       includedPaths.end(),
+                       std::back_inserter(onLocations),
+                       [this](const SdfPath& path)
+                       { return UsdKatanaUtils::ConvertUsdPathToKatLocation(path, _usdInArgs); });
+    }
+    else
+    {
+        const UsdCollectionAPI::MembershipQuery::PathExpansionRuleMap& linkMap{
+            query.GetAsPathExpansionRuleMap()};
+        for (const auto& [path, token] : linkMap)
+        {
+            if (path == SdfPath::AbsoluteRootPath())
+            {
                 // Skip property paths
                 continue;
             }
             const std::string location =
-                UsdKatanaUtils::ConvertUsdPathToKatLocation(entry.first, _usdInArgs);
-            const bool on = (entry.second != UsdTokens->exclude);
-            (on ? onLocations : offLocations).push_back(location);
-            isLinked = true;
+                UsdKatanaUtils::ConvertUsdPathToKatLocation(path, _usdInArgs);
+            const bool isIncluded = (token != UsdTokens->exclude);
+            (isIncluded ? onLocations : offLocations).push_back(location);
         }
     }
 
@@ -2327,11 +2341,18 @@ bool UsdKatanaUtilsLightListAccess::SetLinks(const UsdCollectionAPI& collectionA
 
     if (!onLocations.empty() || !offLocations.empty())
     {
-        std::string onStr = onLocations.empty() ? "" : ConvertVectorToString(onLocations);
+        const std::string onStr = onLocations.empty() ? "" : ConvertVectorToString(onLocations);
         _Set("linking." + linkName + ".onCEL", FnAttribute::StringAttribute(onStr));
 
-        std::string offStr = offLocations.empty() ? "" : ConvertVectorToString(offLocations);
+        const std::string offStr = offLocations.empty() ? "" : ConvertVectorToString(offLocations);
         _Set("linking." + linkName + ".offCEL", FnAttribute::StringAttribute(offStr));
+        _Set("linking." + linkName + ".clearUnmatched", FnAttribute::IntAttribute(0));
+        if (prim.IsA<UsdLuxLightFilter>())
+        {
+            _Set("linking.enable.onCEL", FnAttribute::StringAttribute(onStr));
+            _Set("linking.enable.offCEL", FnAttribute::StringAttribute(offStr));
+            _Set("linking.enable.clearUnmatched", FnAttribute::IntAttribute(0));
+        }
     }
 
     return isLinked;